Fixes #25: refactoring
