+++ pom.xml
@@ -27,12 +27,6 @@
 	<name>api-gateway</name>
 	<description>API Gateway using Feign and Hystrix</description>
 
-	<parent>
-		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.4.4.RELEASE</version>
-	</parent>
-
 	<properties>
 		<!-- Explicitly declaring the source encoding eliminates the following
 			message: -->
@@ -44,10 +38,9 @@
 		<maven.compiler.source>1.8</maven.compiler.source>
 
 		<!-- other versions -->
-		<netflix.feign.version>8.15.1</netflix.feign.version>
-		<hystrix.version>1.5.1</hystrix.version>
+		<camel.version>2.18.2</camel.version>
+		<spring-boot.version>1.4.4.RELEASE</spring-boot.version>
 		<brave.version>4.0.6</brave.version>
-		<swagger.version>1.5.8</swagger.version>
 		<springfox.version>2.4.0</springfox.version>
 
 		<!-- fabric8 version -->
@@ -79,15 +72,18 @@
 		<fabric8.readinessProbe.initialDelaySeconds>5</fabric8.readinessProbe.initialDelaySeconds>
 		<fabric8.readinessProbe.timeoutSeconds>30</fabric8.readinessProbe.timeoutSeconds>
 
-
-		<!-- NEW -->
-		<camel.version>2.18.2</camel.version>
-
 	</properties>
 
 
 	<dependencyManagement>
 		<dependencies>
+			<dependency>
+				<groupId>org.springframework.boot</groupId>
+				<artifactId>spring-boot-dependencies</artifactId>
+				<version>${spring-boot.version}</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
 			<dependency>
 				<groupId>org.apache.camel</groupId>
 				<artifactId>camel-spring-boot-dependencies</artifactId>
@@ -125,51 +121,14 @@
 			<artifactId>camel-swagger-java-starter</artifactId>
 		</dependency>
 
-		<!--<dependency>-->
-			<!--<groupId>org.apache.camel</groupId>-->
-			<!--<artifactId>camel-zipkin-starter</artifactId>-->
-		<!--</dependency>-->
-
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-actuator</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>com.netflix.feign</groupId>
-			<artifactId>feign-core</artifactId>
-			<version>${netflix.feign.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>com.netflix.feign</groupId>
-			<artifactId>feign-hystrix</artifactId>
-			<version>${netflix.feign.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>com.netflix.feign</groupId>
-			<artifactId>feign-jackson</artifactId>
-			<version>${netflix.feign.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>com.netflix.feign</groupId>
-			<artifactId>feign-httpclient</artifactId>
-			<version>${netflix.feign.version}</version>
-		</dependency>
-
-		<!-- hystrix-metrics-event-stream -->
-		<dependency>
-			<groupId>com.netflix.hystrix</groupId>
-			<artifactId>hystrix-core</artifactId>
-			<version>${hystrix.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>com.netflix.hystrix</groupId>
-			<artifactId>hystrix-metrics-event-stream</artifactId>
-			<version>${hystrix.version}</version>
 		</dependency>
 
 		<!-- Brave Zipkin -->
@@ -188,18 +147,8 @@
 			<artifactId>brave-spancollector-http</artifactId>
 			<version>${brave.version}</version>
 		</dependency>
-		<dependency>
-			<groupId>org.apache.httpcomponents</groupId>
-			<artifactId>httpclient</artifactId>
-		</dependency>
 
 		<!-- Swagger -->
-		<dependency>
-			<groupId>io.swagger</groupId>
-			<artifactId>swagger-jaxrs</artifactId>
-			<version>${swagger.version}</version>
-		</dependency>
-
 		<dependency>
 			<groupId>io.springfox</groupId>
 			<artifactId>springfox-swagger2</artifactId>
@@ -219,7 +168,17 @@
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
+				<version>${spring-boot.version}</version>
+				<executions>
+					<execution>
+						<goals>
+							<goal>repackage</goal>
+						</goals>
+					</execution>
+				</executions>
 			</plugin>
+
+
 			<plugin>
 				<groupId>io.fabric8</groupId>
 				<artifactId>fabric8-maven-plugin</artifactId>
