+++ src/main/java/com/redhat/developers/msa/api_gateway/CamelConfiguration.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2016 Red Hat, Inc.
+ *
+ * Red Hat licenses this file to you under the Apache License, version
+ * 2.0 (the "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied.  See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+package com.redhat.developers.msa.api_gateway;
+
+import org.apache.camel.component.hystrix.metrics.servlet.HystrixEventStreamServlet;
+import org.apache.camel.component.servlet.CamelHttpTransportServlet;
+import org.springframework.boot.web.servlet.ServletRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class CamelConfiguration {
+
+    /**
+     * Bind the Camel servlet at the "/api" context path.
+     */
+    @Bean
+    public ServletRegistrationBean camelServletRegistrationBean() {
+        ServletRegistrationBean mapping = new ServletRegistrationBean();
+        mapping.setServlet(new CamelHttpTransportServlet());
+        mapping.addUrlMappings("/api/*");
+        mapping.setName("CamelServlet");
+        mapping.setLoadOnStartup(1);
+
+        return mapping;
+    }
+
+    /**
+     * Bind the Hystrix servlet to /hystrix.stream
+     */
+    @Bean
+    public ServletRegistrationBean hystrixServletRegistrationBean() {
+        ServletRegistrationBean mapping = new ServletRegistrationBean();
+        mapping.setServlet(new HystrixEventStreamServlet());
+        mapping.addUrlMappings("/hystrix.stream");
+        mapping.setName("HystrixEventStreamServlet");
+
+        return mapping;
+    }
+
+}
